requirements.txt
beautifulsoup4
python-dotenv
schedule
emoji
numpy
discord.py
pymongo
pymongo[srv]